refactor: extract starlark conversion utilities to shared file
Moved goToStarlark() and starlarkToGo() helper functions from module_json.go
and module_meow.go to a new converters.go file. These utilities are now shared
across all parsing modules (JSON, YAML, XML, TOML, CSV).

This fixes build errors in XML and TOML modules which were using these functions
but didn't have access to them.
